refactor(api): extract enrichment terminal service from worker
Behavior-preserving extraction of the terminal state machine (succeeded
write, rate-limit deferral, exponential retry, safeTerminalUpdate) from
EnrichmentJobWorker.processJob into EnrichmentTerminalService, so
node-reported results/failures can share the exact same semantics.
Throttle recordSuccess/trip move with it (keyed by
ProviderThrottleService.resolveKey); acquire stays in the worker.
ENRICHMENT_MAX_ATTEMPTS moves to the terminal service and is re-exported
from the worker file for existing consumers. Worker spec adapted to
provide the terminal service and target safeTerminalUpdate at its new
home; all worker spec tests green. Note: 2 pre-existing failures in
enrichment-admin.service.spec.ts (retry reset fields) also fail on
clean HEAD — unrelated.
