+++ apps/api/src/enrichment/enrichment.module.ts
@@ -3,6 +3,7 @@ import { EnrichmentHandlerRegistry } from './enrichment-handler.registry';
 import { EnrichmentJobService } from './enrichment-job.service';
 import { EnrichmentJobWorker } from './enrichment-job.worker';
 import { EnrichmentClaimService } from './enrichment-claim.service';
+import { EnrichmentTerminalService } from './enrichment-terminal.service';
 import { EnrichmentAdminService } from './enrichment-admin.service';
 import { EnrichmentAdminController } from './enrichment-admin.controller';
 import { EnrichmentStuckResetTask } from './enrichment-stuck-reset.task';
@@ -21,6 +22,7 @@ import { SettingsModule } from '../settings/settings.module';
     EnrichmentJobService,
     EnrichmentJobWorker,
     EnrichmentClaimService,
+    EnrichmentTerminalService,
     EnrichmentAdminService,
     EnrichmentStuckResetTask,
     TempFileJanitorTask,
@@ -32,6 +34,7 @@ import { SettingsModule } from '../settings/settings.module';
     EnrichmentJobService,
     EnrichmentHandlerRegistry,
     EnrichmentClaimService,
+    EnrichmentTerminalService,
     EnrichmentAdminService,
   ],
 })
