+++ apps/api/src/enrichment/index.ts
@@ -2,4 +2,5 @@ export * from './enrichment-handler.interface';
 export * from './enrichment-handler.registry';
 export * from './enrichment-job.service';
 export * from './enrichment-job.worker';
+export * from './enrichment-terminal.service';
 export * from './enrichment.module';
